Extract `Asset` and `Ledger` into separate modules
Use enums instead of string-based types.
